+++ AppleALC/kern_alc.cpp
@@ -458,16 +458,16 @@ void AlcEnabler::processKext(KernelPatcher &patcher, size_t index, mach_vm_addre
 	if (progressState & ProcessingState::ControllersLoaded) {
 		for (size_t i = 0, num = controllers.size(); i < num; i++) {
 			auto &info = controllers[i]->info;
-			if (!info) {
+			if (info.size() == 0) {
 				DBGLOG("alc", "missing ControllerModInfo for %lu controller", i);
 				continue;
 			}
 
-			DBGLOG("alc", "handling %lu controller %X:%X with %lu patches", i, info->vendor, info->device, info->patchNum);
+			DBGLOG("alc", "handling %lu controller %X:%X with %lu sets, %lu patches - %s", i, info[0]->vendor, info[0]->device, info.size(), info[0]->patchNum, info[0]->name);
 			// Choose a free device-id for NVIDIA HDAU to support multigpu setups
-			if (info->vendor == WIOKit::VendorID::NVIDIA) {
-				for (size_t j = 0; j < info->patchNum; j++) {
-					auto &p = info->patches[j].patch;
+			if (info[0]->vendor == WIOKit::VendorID::NVIDIA) {
+				for (size_t j = 0; j < info[0]->patchNum; j++) {
+					auto &p = info[0]->patches[j].patch;
 					if (p.size == sizeof(uint32_t) && *reinterpret_cast<const uint32_t *>(p.find) == NvidiaSpecialFind) {
 						DBGLOG("alc", "finding %08X repl at %lu curr %lu", *reinterpret_cast<const uint32_t *>(p.replace), i, currentFreeNvidiaDeviceId);
 						while (currentFreeNvidiaDeviceId < MaxNvidiaDeviceIds) {
@@ -488,7 +488,8 @@ void AlcEnabler::processKext(KernelPatcher &patcher, size_t index, mach_vm_addre
 				DBGLOG("alc", "skipping %lu controller %X:%X:%X due to no-controller-patch", i, controllers[i]->vendor, controllers[i]->device, controllers[i]->revision);
 				continue;
 			}
-			applyPatches(patcher, index, info->patches, info->patchNum);
+			for (size_t i = 0, num = info.size(); i < num; i++)
+				applyPatches(patcher, index, info[i]->patches, info[i]->patchNum);
 		}
 
 		// Only do this if -alcdbg is not passed
@@ -593,21 +594,20 @@ void AlcEnabler::validateControllers() {
 				// Check AAPL,ig-platform-id if present
 				if (ADDPR(controllerMod)[mod].platform != ControllerModInfo::PlatformAny &&
 					ADDPR(controllerMod)[mod].platform != controllers[i]->platform) {
-					DBGLOG("alc", "not matching platform was found %X vs %X", ADDPR(controllerMod)[mod].platform, controllers[i]->platform);
+					DBGLOG("alc", "not matching platform was found %X vs %X for %s", ADDPR(controllerMod)[mod].platform, controllers[i]->platform, &ADDPR(controllerMod)[mod].name);
 					continue;
 				}
 
 				// Check if computer model is suitable
 				if (!(computerModel & ADDPR(controllerMod)[mod].computerModel)) {
-					DBGLOG("alc", "unsuitable computer model was found %X vs %X", ADDPR(controllerMod)[mod].computerModel, computerModel);
+					DBGLOG("alc", "unsuitable computer model was found %X vs %X for %s", ADDPR(controllerMod)[mod].computerModel, computerModel, &ADDPR(controllerMod)[mod].name);
 					continue;
 				}
 
 				if (rev != ADDPR(controllerMod)[mod].revisionNum ||
 					ADDPR(controllerMod)[mod].revisionNum == 0) {
-					DBGLOG("alc", "found mod for %lu controller", i);
-					controllers[i]->info = &ADDPR(controllerMod)[mod];
-					break;
+					DBGLOG("alc", "found mod for %lu controller - %s", i, &ADDPR(controllerMod)[mod].name);
+					controllers[i]->info.push_back(&ADDPR(controllerMod)[mod]);
 				}
 			}
 		}